spark: publish ProcessingEngineRunFacet (#2089)
* spark: publish ProcessingEngineRunFacet

Previously, the Spark integration published a custom facet named
'SparkVersion'. As it is custom, it isn't defined in the OpenLineage
spec. However, the OpenLineage spec defines a ProcessingEngineRunFacet
that is meant to capture details about the things that runs a job.

This change introduces this in the form of the
'SparkProcessingEngineRunFacetBuilder' and
'SparkProcessingEngineRunFacetBuilderDelegate'.

As the names suggest, these classes are meant to create and populate
ProcessingEngineRunFacet.

The reason for the existence of the delegate is because there are two
code paths that interact with the run facets, namely the code path
within the RddExecutionContext and the other in the
SparkSqlExecutionContext, albeit via a very roundabout way.

The delegate is the object that actually constructs the facet, whilst
the builder provides an adapter that uses the CustomFacetBuilder
interface.

Yes, it's a hacky way of doing it and may need to be changed in the
future. For now though, its good enough.

Closes: https://github.com/OpenLineage/OpenLineage/issues/2086
Signed-off-by: Damien Hawes <d-m-h@users.noreply.github.com>

* spark: Deprecated the SparkVersionFacet, alongside the version-facet.json

Signed-off-by: Damien Hawes <d-m-h@users.noreply.github.com>

* spark: Used a mocked SparkContext instead inside SparkProcessingEngineFacetBuilderTest

Signed-off-by: Damien Hawes <d-m-h@users.noreply.github.com>

* changelog: Updated the changelog indicating that the SparkVersionFacet will be removed in 1.4.0

Signed-off-by: Damien Hawes <d-m-h@users.noreply.github.com>

---------

Signed-off-by: Damien Hawes <d-m-h@users.noreply.github.com>