Fix BigQuery symlinks with ".db" suffix (#4192)
".db" suffix is normally added to a namespace in context of
physical location. When working with logical location
(which is what we can find in symlink), there should be
no such suffix.
The last element of namespace array was being modified
in place (mutated) to add a ".db" suffix.
Since the same namespace array from the Spark Identifier object
is used both to create the logical symlink (via getIdentifier())
and to construct the physical path (via defaultTableLocation()),
this mutation could cause the ".db" suffix to leak into
the symlink namespace where it shouldn't appear.

Signed-off-by: kchledowski <github@chledowski.com>